# contracts/api/books/v1/openapi.yaml
openapi: 3.0.3
info:
  title: LoreWeave Books API
  version: 1.5.0
  description: |
    Module 02 — books, chapters (MVP text/plain .txt uploads), editable chapter draft (UTF-8) with revision history,
    cover image, summary, and per-user storage usage. Multilingual: each locale is a separate chapter row with
    original_language; list chapters supports optional query filters original_language and/or sort_order and
    pagination (limit/offset/total). Owner list includes visibility for direct badge rendering in workspace.
    Logical delete: DELETE book/chapter moves to recycle bin (trashed); DELETE …/purge from bin marks purge_pending for
    future garbage collection (blobs + DB rows removed later by a batch job, not in this API). GET /v1/books lists active
    books only; GET /v1/books/trash lists trashed books. Trashed or purge_pending books must not appear in catalog/unlisted/public.
    Raw upload bytes remain in S3-compatible storage (e.g. MinIO); draft/revisions in Postgres until purged. Gateway-facing REST only.
servers:
  - url: http://localhost:3000
    description: API Gateway (BFF)
...
paths:
...
  /v1/import-jobs/{job_id}/rollback:
    post:
      tags: [imports]
      summary: Roll back the job-owned effects of an EPUB import
      operationId: rollbackEpubImportJob
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ImportJobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties:
                confirm:
                  type: boolean
                  enum: [true]
                  description: Explicit acknowledgement that job-owned import effects will be removed or deactivated.
      responses:
        '202':
          description: Rollback accepted or was already completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpubImportJobActionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ImportJobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ErrorResponse:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    EpubImportJobStatus:
      type: string
      enum:
        - queued
        - processing
        - import_staging
        - import_ready
        - completed
        - completed_with_warnings
        - import_failed
        - cancelled
        - rolling_back
        - rolled_back
...
    EpubImportJobActionResponse:
      type: object
      required: [job_id, status]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/EpubImportJobStatus'
        chapters_rolled_back:
          type: integer
          minimum: 0
          description: Present for rollback responses; counts chapters safely removed by this call.
        conflicts:
          type: array
          description: Chapters retained because they changed after import finalization.
          items:
            type: object
            required: [chapter_id, code]
            properties:
              chapter_id:
                type: string
                format: uuid
              code:
                type: string
                enum: [rollback_conflict_user_modified]
...